change qty on carts table to integer

+++ database/migrations/2022_02_15_144944_create_carts_table.php
@@ -17,7 +17,7 @@ class CreateCartsTable extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained();
             $table->foreignId('product_id')->constrained();
-            $table->float('qty')->nullable();
+            $table->unsignedBigInteger('qty')->nullable();
             $table->timestamps();
         });
     }
